requirements.txt
# Celery Packages
celery==4.4.0
flower==0.9.3
kombu==4.6.7

# Databases
psycopg2-binary==2.8.4
SQLAlchemy==1.3.11
sqlalchemy-utils==0.35.0
redis==3.3.11
alembic==1.3.2

# Flask
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-json-schema==0.0.5
Flask-Migrate==2.5.2

# Misc
gunicorn==20.0.4
environs==6.1.0
pyyaml==5.1.2
click==7.0
terminaltables==3.1.0
pluginbase==1.0.0
